+++ audiences-react/src/AudienceForm/index.tsx
@@ -1,61 +1,32 @@
 import { useState } from "react"
-import { FixedConfirmationToast, Button, Flex } from "playbook-ui"
+import { FixedConfirmationToast, Button, FlexItem } from "playbook-ui"
+import { isEmpty } from "lodash/fp"
 
-import { GroupCriterion, ScimObject } from "../types"
+import { ScimObject } from "../types"
 import { toSentence } from "./toSentence"
 
 import { ScimResourceTypeahead } from "./ScimResourceTypeahead"
 import { CriteriaList } from "./CriteriaList"
 import { CriteriaForm } from "./CriteriaForm"
-import { useAudiences } from "../audiences"
+import { ActionBar } from "./ActionBar"
 import { MatchAllToggleCard } from "./MatchAllToggleHeader"
+import { useAudiencesContext } from "../audiences"
 
 type AudienceFormProps = {
-  uri: string
   userResource: string
   groupResources: string[]
   allowIndividuals: boolean
-  fetchOptions: Parameters<typeof useAudiences>[1]
 }
 
 export const AudienceForm = ({
-  uri,
   userResource,
   groupResources,
   allowIndividuals = true,
-  fetchOptions,
 }: AudienceFormProps) => {
   const [editing, setEditing] = useState<number>()
+  const { error, value: context, change } = useAudiencesContext()
 
-  const {
-    saving,
-    fetchUsers,
-    save,
-    error,
-    value: context,
-    isDirty,
-    change,
-    reset,
-    removeCriteria,
-    updateCriteria,
-  } = useAudiences(uri, fetchOptions)
-
-  const handleRemoveCriteria = (index: number) => {
-    if (confirm("Remove criteria?")) {
-      removeCriteria(index)
-    }
-  }
-
-  const handleSaveCriteria = (criterion: GroupCriterion) => {
-    updateCriteria(editing!, criterion)
-    setEditing(undefined)
-  }
-
-  const handleCreateCriteria = () => {
-    setEditing(context.criteria.length)
-  }
-
-  if (!context) {
+  if (isEmpty(context)) {
     return null
   }
 
@@ -63,62 +34,41 @@ export const AudienceForm = ({
     return (
       <CriteriaForm
         resources={groupResources}
-        criterion={context.criteria[editing]}
-        onSave={handleSaveCriteria}
-        onCancel={() => setEditing(undefined)}
+        criterion={editing}
+        onExit={() => setEditing(undefined)}
       />
     )
   }
 
   return (
-    <MatchAllToggleCard
-      count={context.count}
-      enabled={context.match_all}
-      fetchUsers={fetchUsers}
-      isDirty={isDirty()}
-      onToggle={(all: boolean) => change("match_all", all)}
-    >
+    <MatchAllToggleCard>
       {error && (
         <FixedConfirmationToast status="error" text={error} margin="sm" />
       )}
-      {allowIndividuals && !context.match_all && (
-        <ScimResourceTypeahead
-          label="Add Individuals"
-          value={context.extra_users || []}
-          onChange={(users: ScimObject[]) => change("extra_users", users)}
-          resourceId={userResource}
-        />
-      )}
       {!context.match_all && (
-        <CriteriaList
-          addCriteriaLabel={`Add Members by ${toSentence(groupResources)}`}
-          context={context}
-          fetchUsers={fetchUsers}
-          onAddCriteria={handleCreateCriteria}
-          onEditCriteria={setEditing}
-          onRemoveCriteria={handleRemoveCriteria}
-        />
+        <>
+          {allowIndividuals && (
+            <ScimResourceTypeahead
+              label="Add Individuals"
+              value={context.extra_users || []}
+              onChange={(users: ScimObject[]) => change("extra_users", users)}
+              resourceId={userResource}
+            />
+          )}
+          <CriteriaList onEditCriteria={setEditing} />
+          <FlexItem alignSelf="center">
+            <Button
+              fixedWidth
+              marginTop="md"
+              onClick={() => setEditing(context.criteria.length)}
+              text={`Add Members by ${toSentence(groupResources)}`}
+              variant="link"
+            />
+          </FlexItem>
+        </>
       )}
 
-      <Flex justify="between" marginTop="md">
-        <Button
-          disabled={!isDirty}
-          text="Save"
-          htmlType="submit"
-          loading={saving}
-          onClick={() => save()}
-        />
-
-        {isDirty() && (
-          <Button
-            marginLeft="sm"
-            text="Cancel"
-            variant="link"
-            htmlType="reset"
-            onClick={() => reset()}
-          />
-        )}
-      </Flex>
+      <ActionBar />
     </MatchAllToggleCard>
   )
 }